Negative speed and ledHandler Module
The Motor Libary support negative speed and the LedHandler is now in a seperate file.

+++ Homekit codes/Homekit Modules/motor_speed_and_led/src/mapper/mapper.h
@@ -5,13 +5,13 @@ class Map
 {
   // A class that linearly maps between input range of values to output range of values
 public:
-  float input_min_value;
-  float input_max_value;
-  float Output_min_value;
-  float output_max_value;
-  float input_scale_size;
-  float output_scale_size;
-  Map(float input_min_value, float input_max_value, float output_min_value, float output_max_value);
-  float map_value(float input_value);
+  int input_min_value;
+  int input_max_value;
+  int Output_min_value;
+  int output_max_value;
+  int input_scale_size;
+  int output_scale_size;
+  Map(int input_min_value, int input_max_value, int output_min_value, int output_max_value);
+  int map_value(int input_value);
 };
 #endif
